python/agent-hub/MemoryAugmentedChatAssistant/pyproject.toml
[tool.poetry]
name = "MemoryAugmentedChatAssistant"
version = "0.1.0"
description = "Auto-generated agent package"
authors = [
    "Mofa Bot <mofa-bot@moxin.com>",
    "Technical Owner <eng@moxin.com>"
]
packages = [{ include = "memory_augmented_chat" }]

[tool.poetry.dependencies]
python = ">=3.10"
pyarrow = ">= 5.0.0"
openai = "*"
mem0ai = "*"

[tool.poetry.scripts]
MemoryAugmentedChatAssistant = "memory_augmented_chat.main:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
